rename filter_ext module to retain_ext
